requirements.txt
Flask==2.1.0
numpy==1.21.2
Pillow==8.3.1
requests==2.26.0
streamlit==1.15.0